Rename _nucleotide_sequence.py
Changed to _nucleotide_mixin.py
