+++ skbio/sequence/_nucleotide_mixin.py
